+++ booquest-api/src/test/java/com/booquest/booquest_api/application/service/onboarding/OnboardingServiceTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
@@ -12,7 +11,7 @@ import static org.mockito.Mockito.verifyNoMoreInteractions;
 import com.booquest.booquest_api.application.port.in.dto.SubmitOnboardingData;
 import com.booquest.booquest_api.application.port.out.onboarding.OnboardingCategoryRepository;
 import com.booquest.booquest_api.application.port.out.onboarding.OnboardingProfileRepository;
-import com.booquest.booquest_api.adapter.out.auth.persistence.jpa.UserRepository;
+import com.booquest.booquest_api.adapter.out.user.persistence.UserRepository;
 import com.booquest.booquest_api.domain.onboarding.enums.ExpressionStyle;
 import com.booquest.booquest_api.domain.onboarding.enums.StrengthType;
 import com.booquest.booquest_api.domain.onboarding.model.OnboardingProfile;
